Add support for Silverstripe 5
