+++ composer.json
@@ -15,10 +15,11 @@
         "issues": "https://github.com/kinglozzer/silverstripe-metatitle/issues"
     },
     "require": {
-        "silverstripe/cms": "^4"
+        "php": "^8.1",
+        "silverstripe/cms": "^5"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.7"
+        "phpunit/phpunit": "^9.5"
     },
     "autoload": {
         "psr-4": {
